更新Resource相关接口

+++ app/controllers/api/v1/users_controller.rb
@@ -1,8 +1,11 @@
 class Api::V1::UsersController < ApplicationController
 
+  before_action :authenticate_user
+
   def index
     @users = User.with_search(params).with_sort(params)
     @users = @users.page(params[:page] || 1).per(params[:per])
+        .includes(:user_info)
     render json: @users, root: 'items', meta: meta_with_page(@users)
   end
 
@@ -24,6 +27,7 @@ class Api::V1::UsersController < ApplicationController
 
   def update
     @user = User.find(params[:id])
+    authorize @user, :update_or_destroy?
     @user.update_user_info(params)
     if @user.update!(user_params)
       render json: @user
@@ -34,6 +38,7 @@ class Api::V1::UsersController < ApplicationController
 
   def destroy
     @user = User.find(params[:id])
+    authorize @user, :update_or_destroy?
     if @user.destroy
       render json: { error_code: 0 }
     else
@@ -41,9 +46,10 @@ class Api::V1::UsersController < ApplicationController
     end
   end
 
+  private
+
   def user_params
     params.permit(:name, :nickname, :gender, :avatar, :role, :status, :description)
   end
 
-  private :user_params
 end
